Use indexed triangles for submit
Use a very messy quick-n-dirty implementation of indexed triangles
in the example to help find a pattern to build upon moving forward

Thoughts so far:

Maintaining the number of indicies between functions which generate
verts is annoying when the calls are dispirate, especially if I try
not to pass around mutable values.

I wonder if it's better to use some sort of simple abstraction for
primitives that get resolved to indexed triangles at a defined stage
or if there should be some sort of state passed around which
"renders" indexed triangles at every function call ala Dear ImGui.
Perhaps both approaches are worth investigating.