Experiment with static INDICIES counter
Incrementing the `INDICIES` counter by the number of verts put out
during the `rect` function during the function itself is nice in
that I don't have to keep track of the number of indicies anywhere
in the program but the areas it matters; however, it is unsafe
and would probably not be thread friendly; I could try using an
atomic type since it's just a number, but there's another issue
as well: the counter must be reset to zero.  So I assume in Dear,
ImGui it's reset at some point when you retrieve the verts from
the API and I would have to do something similar.  It works for
Dear, ImGui's programming style but I feel like I'd rather work
in an environment with parameter passing.

Next I will try some sort of system of primitives which get
rendered out in a dedicated phase, perhaps even with some
sort of callback function which actually accepts the vertex
data and consumes it how it pleases.