Add a polyline implementation
Polyline implementation supports miter'd polylines which are joined
together rather than drawn as individual segments.

Includes a benchmark.

Implementation thanks to:
https://forum.libcinder.org/topic/smooth-thick-lines-using-geometry-shader

Polylines are not anti-aliased (yet) and are not really optimized
although I did add a benchmark.

Will compare performance to an unjoined polyline implementation and
will consider which to use as the default.