Create top level structure
Created a UI struct and several other types which:

- defines the top-level context of the UI
- a separate interface for accessing render data
- a few methods to keep track of active items

The active tracking is completely non-functional but was thrown in
just to have something to build upon.  Future work will include:

- designing the API for setting an element as active based on
  events
- Querying for basic click and keyboard events on the active item
  in a specific region

And more.  I have a lot of thoughts about how this should work but
they aren't really coherently collected at the moment.  I am hoping
that by getting something working and iterating I can clarify those
thoughts.

I would like to create a uniform system for keeping track of simple
data between frames, something similar to the react hooks or the
moxie-rs memoization store, but with a limited scope.

Or maybe if that doesn't work out, I'll take an approach more like
gioui which requires that any state be managed by the application.

But for simple things like keeping track of a bool for toggles or
a float for scroll bars (in things that can't be derived from user
data) I'm more inclined to let the UI handle that.

We'll see how this evolves!