Use Option<Vec2> for input mouse_pos
The mouse position can be off the screen or sometimes there is no
useful value for the mouse position; therefore it makes sense, at
least for the moment, to provide no mouse position initially.

It may not make much sense in practice, and using a mouse position
outside of screen coordinates may make more sense in the future.